updated type to float

+++ src/odrive_can_node.cpp
@@ -315,8 +315,8 @@ void ODriveCanNode::control_gains_callback(const odrive_can::msg::ControlGains::
     frame.can_id = node_id_ << 5 | kSetVelGains;
     {
         std::lock_guard<std::mutex> guard(gains_msg_mutex_);
-        write_le<uint32_t>(msg->vel_gain, frame.data);
-        write_le<uint32_t>(msg->vel_integrator_gain,   frame.data + 4);
+        write_le<float>(msg->vel_gain, frame.data);
+        write_le<float>(msg->vel_integrator_gain,   frame.data + 4);
     }
     frame.can_dlc = 8;
     can_intf_.send_can_frame(frame);
